Set up build, add basic files
LICENSE is from the JSON Schema spec repo, replacing
"JSON Schema Specification Authors" with
"JSON Schema Organization".

CONTRIBUTING.md is a highly condensed adaptation of
the file in the JSON Schema spec repo.

BACKGROUND.md is mostly stuff that was in appendixes of an
early draft of a proposal that does not need to be part of
a standard at all.  There's too much of it to put in the README,
but I got feedback that the material was very useful for some folks.

Makefile is adapted from the IETF HTTP APIs working group spec
repository.  The only change is to set the CSS to the style
used for published I-Ds.  For some reason, that is not the deafult.

The presence of "lib/" is .gitignore is because that is where
the Makefile checks out the tools and creates a virtual env.